fix(client): fix ecma version check worker error
fix(client): fix ecma version check worker error

fix(client): fix ecma version check worker error
