+++ .changeset/tiny-otters-hide.md
@@ -1,0 +1,5 @@
+---
+'@rsdoctor/components': patch
+---
+
+fix(client): ecma version check worker error
